+++ index.d.ts
@@ -14,6 +14,14 @@ type CamelCase<S extends string> = S extends `${infer W}-${infer Rest}`
 // - https://effectivetypescript.com/2022/02/25/gentips-4-display/
 type Resolve<T> = T extends infer O ? { [K in keyof O]: O[K] } : never;
 
+// This is a trick to encourage editor to suggest the known literals while still
+// allowing any BaseType value.
+// References:
+// - https://github.com/microsoft/TypeScript/issues/29729
+// - https://github.com/sindresorhus/type-fest/blob/main/source/literal-union.d.ts
+// - https://github.com/sindresorhus/type-fest/blob/main/source/primitive.d.ts
+type LiteralUnion<LiteralType, BaseType extends string | number> = LiteralType | (BaseType & Record<never, never>);
+
 // Side note: not trying to represent arrays as non-empty, keep it simple.
 // https://stackoverflow.com/a/56006703/1082434
 type InferVariadic<S extends string, ArgT> =
@@ -72,10 +80,10 @@ type ConvertFlagToName<Flag extends string> =
             ? CamelCase<Name>
             : never;
 
-type CombineOptions<Options extends Record<string, any>, B extends Record<string, any>> =
-    keyof B extends keyof Options
-        ? { [K in keyof Options]: K extends keyof B ? Options[K] | B[keyof B] : Options[K] }
-        : Resolve<Options & B>;
+type CombineOptions<Options, O> =
+    keyof O extends keyof Options
+        ? { [K in keyof Options]: K extends keyof O ? Options[K] | O[keyof O] : Options[K] }
+        : Options & O;
 
 type IsAlwaysDefined<DefaulT, Mandatory extends boolean> =
     Mandatory extends true
@@ -147,6 +155,8 @@ type InferOptions<Options, Usage extends string, DefaultT, CoerceT, Mandatory ex
     : Usage extends `${infer Flags} [${infer Value}]`
       ? InferOptionsFlag<Options, Flags, Value, string, PresetT, DefaultT, CoerceT, Mandatory>
       : InferOptionsFlag<Options, Usage, '', never, PresetT, DefaultT, CoerceT, Mandatory>;
+
+export type CommandUnknownOpts = Command<unknown[], OptionValues>;
 
 // ----- full copy of normal commander typings from here down, with extra type inference -----
 
@@ -293,7 +303,9 @@ export class CommanderError extends Error {
   
     /**
      * Set environment variable to check for option value.
-     * Priority order of option values is default < env < cli
+     *
+     * An environment variables is only used if when processed the current option value is
+     * undefined, or the source of the current value is 'default' or 'config' or 'env'.
      */
     env(name: string): this;
   
@@ -350,9 +362,9 @@ export class CommanderError extends Error {
     constructor();
   
     /** Get the command term to show in the list of subcommands. */
-    subcommandTerm(cmd: Command): string;
+    subcommandTerm(cmd: CommandUnknownOpts): string;
     /** Get the command summary to show in the list of subcommands. */
-    subcommandDescription(cmd: Command): string;
+    subcommandDescription(cmd: CommandUnknownOpts): string;
     /** Get the option term to show in the list of options. */
     optionTerm(option: Option): string;
     /** Get the option description to show in the list of options. */
@@ -363,25 +375,25 @@ export class CommanderError extends Error {
     argumentDescription(argument: Argument): string;
   
     /** Get the command usage to be displayed at the top of the built-in help. */
-    commandUsage(cmd: Command): string;
+    commandUsage(cmd: CommandUnknownOpts): string;
     /** Get the description for the command. */
-    commandDescription(cmd: Command): string;
+    commandDescription(cmd: CommandUnknownOpts): string;
   
     /** Get an array of the visible subcommands. Includes a placeholder for the implicit help command, if there is one. */
-    visibleCommands(cmd: Command): Command[];
+    visibleCommands(cmd: CommandUnknownOpts): CommandUnknownOpts[];
     /** Get an array of the visible options. Includes a placeholder for the implicit help option, if there is one. */
-    visibleOptions(cmd: Command): Option[];
+    visibleOptions(cmd: CommandUnknownOpts): Option[];
     /** Get an array of the arguments which have descriptions. */
-    visibleArguments(cmd: Command): Argument[];
+    visibleArguments(cmd: CommandUnknownOpts): Argument[];
   
     /** Get the longest command term length. */
-    longestSubcommandTermLength(cmd: Command, helper: Help): number;
+    longestSubcommandTermLength(cmd: CommandUnknownOpts, helper: Help): number;
     /** Get the longest option term length. */
-    longestOptionTermLength(cmd: Command, helper: Help): number;
+    longestOptionTermLength(cmd: CommandUnknownOpts, helper: Help): number;
     /** Get the longest argument term length. */
-    longestArgumentTermLength(cmd: Command, helper: Help): number;
+    longestArgumentTermLength(cmd: CommandUnknownOpts, helper: Help): number;
     /** Calculate the pad width from the maximum term length. */
-    padWidth(cmd: Command, helper: Help): number;
+    padWidth(cmd: CommandUnknownOpts, helper: Help): number;
   
     /**
      * Wrap the given string to width characters per line, with lines after the first indented.
@@ -390,7 +402,7 @@ export class CommanderError extends Error {
     wrap(str: string, width: number, indent: number, minColumnWidth?: number): string;
   
     /** Generate the built-in help text. */
-    formatHelp(cmd: Command, helper: Help): string;
+    formatHelp(cmd: CommandUnknownOpts, helper: Help): string;
   }
   export type HelpConfiguration = Partial<Help>;
   
@@ -415,17 +427,18 @@ export class CommanderError extends Error {
   
   export type AddHelpTextPosition = 'beforeAll' | 'before' | 'after' | 'afterAll';
   export type HookEvent = 'preSubcommand' | 'preAction' | 'postAction';
-  export type OptionValueSource = 'default' | 'env' | 'config' | 'cli';
-  
+  // The source is a string so author can define their own too.
+  export type OptionValueSource = LiteralUnion<'default' | 'config' | 'env' | 'cli' | 'implied', string> | undefined;
+
   export interface OptionValues {
-    [key: string]: any;
+    [key: string]: unknown;
   }
   
-  export class Command<Args extends any[] = [], Opts = {}> {
+  export class Command<Args extends any[] = [], Opts extends OptionValues = {}> {
     args: string[];
     processedArgs: any[];
-    commands: Command[];
-    parent: Command | null;
+    commands: CommandUnknownOpts[];
+    parent: CommandUnknownOpts | null;
   
     constructor(name?: string);
   
@@ -495,7 +508,7 @@ export class CommanderError extends Error {
      *
      * @returns `this` command for chaining
      */
-    addCommand(cmd: Command, opts?: CommandOptions): this;
+    addCommand(cmd: CommandUnknownOpts, opts?: CommandOptions): this;
   
     /**
      * Factory routine to create a new unattached argument.
@@ -568,7 +581,7 @@ export class CommanderError extends Error {
     /**
      * Add hook for life cycle event.
      */
-    hook(event: HookEvent, listener: (thisCommand: Command, actionCommand: Command) => void | Promise<void>): this;
+    hook(event: HookEvent, listener: (thisCommand: this, actionCommand: CommandUnknownOpts) => void | Promise<void>): this;
   
     /**
      * Register callback to use as replacement for calling process.exit.
@@ -619,7 +632,7 @@ export class CommanderError extends Error {
      *
      * (Used internally when adding a command using `.command()` so subcommands inherit parent settings.)
      */
-    copyInheritedSettings(sourceCommand: Command): this;
+    copyInheritedSettings(sourceCommand: CommandUnknownOpts): this;
   
     /**
      * Display the help or a custom message after an error occurs.